back/requirements.txt
pytest
flask
flask_corsclear
